feat(collection): new component CollectionActivity
